Create README.md
ore if thouest doth protest of thine relation of thine planatary shaped format, wouldst though think it be circular or plain in form? however the reality of the existence to the entrance of the varience in movement relating to thine other degrees and inter planaletary galactic catacalysms , we would have burned to mere ash at the first pass beyond the gasiest Globe that burns bright and casts its rays upon us. Fore life and us and the entirety of being is related to the circular motion and creations are made in its likeness, just as the nebula is a replica of the birth of a zygote, a eye as the same as a black hole, the roots and of the trees and the placenta that feeds, as is the cells that grow as the flower and more.
we are the future of evolution as 
well as the past. 
We are the dreamers the seekers the wayfinders the ones 
We all Carry too much unbridled energy.
 We live in a society that views differences as 
a flaw.. 
Thinking everyone should fit a mold . 
We are the only animals on the planet that punish each other for what we consume and protecting our own bodies. 
And Star shaped creatures can't fit into tiny circular boxes.
If you forfeit ethics, to cash in for a Payday, forgoing all forthought on 
the consequential effects. 
Your at fault, not the latter
  feeble attempts to feel powerful.
Make your soul f'ugly
 Fore you see ...
Fictitious fallacies 
fundamentally 
frame Families; 
forlorn friends feud;
fists fly.
Fountains of flowing faithfulness fall, 
fanatics fill with fear. From..
Flu shots
and fake exposure notifications
Famine still exists
The world is plagued by flat earthed farts
Unable to grasp a reality farther then 
The famed ferris wheel they return to daily
Just the same motion stagnant
Until everything just begins to ..........

..........

FREEZE .... 

..........

Flee
Find new ways..
Fill your space with flair and funk
FEED THE HUNGRY
FOSTER THE SICK 
AND THOSE WHO ARE WITHOUT
Flood with LOVE
Be a Speaker for the music that leads  
Followers who reach out
Need a teacher
Poetically featured 
Fantastic figurines 
Poised with flaire
Finessed with fairness
Focussed
And fluffy, floppy, and frumpy but never foolish unless for fun

Flowered fluidity
Believe in fairy's 
Be as the feather
And float on 
Be as the Geese, find your fleet
Then fly as a united front and mate for life.  
 
Don't flock to the foolish fraudulent framed faces
Like they did to the Fleur.  
And forgot feelings and Empathy
TEACH our past fully , not flipped or faultered 
TO not just our youth.
 But 
to all bretheren alike.
So we may never forget..
Foolish pride funks up lives
And ensure we don't repeat the
Fudgery! 

Face forward , feet steady
 About Front! 
Foster foundations funded in FACTS...
 !! Remembering to be sympathetic and foster familiarizing yourself with 
empathic reasonings
As to not be apathetic 
To the justified reasons why 
Forgive them if they are not sorry
But dont forgo self respect
And that doesn't make make them immune to a lesson taught back
To keep them on track
 let those that taunt and main
Feel the pain that you endured 
When they locked you in a cage,
 but it's small and you don't fit! 
Look at the past and all of the logistics 
In a flaw
In order to figure out
how to be fair, 
how it feels to be FORGOTTEN. 
Fused with fools, 
Felonious Fiends that fling lies like boomerangs 
disguise faces to hide their flawed,
 quintessential Red Flags 
Fools that can't figure out
 we can smell their male-heffer feces
From fifty five thousand feet away
Use intuition and empathy, not just what you've been spoon fed... 
FROM Greedy and Lustfull fans of sin

 FIND feelers for hidden achives, 
WITHIN you might see falsified financials , 
and free former foes